currencyapiserver/requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.3.4
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
billiard==3.3.0.23
brotlipy==0.7.0
celery @ file:///home/conda/feedstock_root/build_artifacts/celery_1616560202256/work
certifi==2021.5.30
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1613414067184/work
chardet @ file:///Users/runner/miniforge3/conda-bld/chardet_1610093477312/work
click @ file:///Users/runner/miniforge3/conda-bld/click_1621503759047/work
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1616851512278/work
cycler==0.10.0
dj-database-url==0.5.0
Django==3.2.4
django-celery==3.3.1
django-heroku==0.3.1
django-kombu==0.9.4
django-ssl-auth==2.2.0
django-sslify==0.2.7
django-sslserver==0.22
django-widget-tweaks==1.4.8
djangorestframework @ file:///home/conda/feedstock_root/build_artifacts/djangorestframework_1616757768152/work
djangorestframework-jsonapi==4.2.0
escapejson==1.0
gunicorn==20.1.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1593328102638/work
inflection==0.5.1
json-serializer==1.0.1
kiwisolver @ file:///opt/concourse/worker/volumes/live/0b2f3e77-eaa3-4995-7dd0-c994762fcbde/volume/kiwisolver_1612282417472/work
kombu @ file:///Users/runner/miniforge3/conda-bld/kombu_1612809629737/work
llvmlite==0.36.0
matplotlib @ file:///Users/runner/miniforge3/conda-bld/matplotlib-base_1594091691585/work
mkl-fft==1.3.0
mkl-random @ file:///opt/concourse/worker/volumes/live/54b31a45-1da5-4512-5c3a-93c9ff2af8bc/volume/mkl_random_1618853970587/work
mkl-service==2.3.0
numba @ file:///opt/concourse/worker/volumes/live/263a950e-7ddc-4297-63df-0c284f0b6d22/volume/numba_1616774255536/work
numpy @ file:///opt/concourse/worker/volumes/live/84716fc3-63c0-4ad6-5a23-37f5ad6b3bfa/volume/numpy_and_numpy_base_1618497240837/work
olefile==0.46
opencv-python==4.5.1.48
pandas==1.2.4
Pillow @ file:///opt/concourse/worker/volumes/live/ca23594b-6e35-4c8c-5637-50ac0b550473/volume/pillow_1617386168018/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1616432837031/work
psycopg2 @ file:///opt/concourse/worker/volumes/live/c8a4b965-959f-4e6c-7792-4aa881d63c8d/volume/psycopg2_1594305121341/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
pyparsing==2.4.7
PySocks @ file:///Users/runner/miniforge3/conda-bld/pysocks_1610291493748/work
python-dateutil==2.8.1
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1608156231189/work
six @ file:///opt/concourse/worker/volumes/live/5b31cb27-1e37-4ca5-6e9f-86246eb206d2/volume/six_1605205320872/work
SQLAlchemy @ file:///Users/runner/miniforge3/conda-bld/sqlalchemy_1612225115309/work
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1610094711612/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1615828766818/work
vine @ file:///home/conda/feedstock_root/build_artifacts/vine_1604930048994/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
whitenoise==5.2.0
